Refactor chemistry_agent to modular directory structure
- Move chemistry_agent.py to directory structure (chemistry/)
- Split into separate modules: chemistry_agent.py, chemistry_config.py, chemistry_tools.py, chemistry_gradio.py, image_agent.py
- Fix circular import issues by separating agent and UI code
- Update CLI to include chemistry command
- Update test imports to use new module structure
- Make tool functions async for better performance
- Follow the same pattern as LinkML and Robot agents
