+++ src/aurelian/agents/chemistry/__init__.py
@@ -1,0 +1,3 @@
+"""
+Chemistry agent module for working with chemical structures.
+"""
